Промежуточный коммит

+++ src/main/java/com/example/cloniamix/weatherapp/weatherApi/POJO/Wind.java
@@ -10,7 +10,7 @@ public class Wind {
     private Double speed;
     @SerializedName("deg")
     @Expose
-    private Integer deg;
+    private double deg;
 
     public Double getSpeed() {
         return speed;
@@ -20,11 +20,11 @@ public class Wind {
         this.speed = speed;
     }
 
-    public Integer getDeg() {
+    public double getDeg() {
         return deg;
     }
 
-    public void setDeg(Integer deg) {
+    public void setDeg(double deg) {
         this.deg = deg;
     }
